ids are now Long type

+++ src/edu/ufp/inf/en/siu/Node.java
@@ -7,7 +7,7 @@ import edu.ufp.inf.en.lp2._1_intro.geometric_figures.Point;
 
 public class Node {
 
-  private Integer nodeId; 
+  private Long nodeId; 
   private Point coordinates;
   private SeparateChainingHashST<Tag, String> tags = new SeparateChainingHashST<>();
 
@@ -15,16 +15,16 @@ public class Node {
     
   }
 
-  public Node (Integer nodeId, Point coordinates) {
+  public Node (Long nodeId, Point coordinates) {
     this();
     this.nodeId = nodeId;
     this.coordinates = coordinates;
   }
 
-  public Integer getNodeId() {
+  public Long getNodeId() {
     return nodeId;
   }
-  public void setNodeId(Integer nodeId) {
+  public void setNodeId(Long nodeId) {
     this.nodeId = nodeId;
   }
   public Point getCoordinates() {
